% Generated by roxygen2 (4.0.1): do not edit by hand
\name{knn1}
\alias{knn1}
\title{Find K nearest neighbours for a single query point}
\usage{
knn1(M, q, k, eps = 0)
}
\arguments{
\item{M}{dxM matrix of M target points with dimension d}

\item{q}{a length d vector defining a query point}

\item{k}{an integer number of nearest neighbours to find}

\item{eps}{An approximate error bound. The default of 0 implies exact
matching.}
}
\value{
A list with elements \code{indices} (1-indexed indices) and
  \code{dists} (distances)
}
\description{
Find K nearest neighbours for a single query point
}
\details{
note that libnabo returns squared distances by default, but we
  unsquare them.
}

