% Generated by roxygen2 (4.0.1): do not edit by hand
\name{knn}
\alias{knn}
\title{Find K nearest neighbours for multiple query points}
\usage{
knn(M, q, k, eps = 0)
}
\arguments{
\item{M}{dxM matrix of M target points with dimension d}

\item{q}{dxN matrix of N query points with dimension d (nb \code{M} and
\code{q} must have same dimension)}

\item{k}{an integer number of nearest neighbours to find}

\item{eps}{An approximate error bound. The default of 0 implies exact
matching.}
}
\value{
A list with elements \code{indices} (1-indexed indices) and
  \code{dists} (distances), both of which are k x N matrices
}
\description{
Find K nearest neighbours for multiple query points
}
\details{
note that libnabo returns squared distances by default, but we
  unsquare them.
}

