% Generated by roxygen2 (4.0.1): do not edit by hand
\name{knn}
\alias{knn}
\alias{knn_brute}
\title{Find K nearest neighbours for multiple query points}
\usage{
knn(data, query, k, eps = 0)

knn_brute(data, query, k)
}
\arguments{
\item{data}{Mxd matrix of M target points with dimension d}

\item{query}{Nxd matrix of N query points with dimension d (nb \code{data} and
\code{query} must have same dimension)}

\item{k}{an integer number of nearest neighbours to find}

\item{eps}{An approximate error bound. The default of 0 implies exact
matching.}
}
\value{
A list with elements \code{nn.idx} (1-indexed indices) and
  \code{nn.dists} (distances), both of which are N x k matrices
}
\description{
knn uses a k-d tree optimised for k ~< 30 nearest neighbours

knn_brute checks all point combinations (for validation only)
}

